refactor(store): extract plugins domain into a zustand slice
Move the plugins domain out of the monolithic appStore.ts into a
dedicated StateCreator slice at src/store/slices/pluginsSlice.ts,
combined into the root create() via spread — mirroring the SSH tunnel
(#2077), embedded-servers (#2113), and macros (#2114) slices.

The slice owns the installed-plugin list plus the backend-type and
theme registries derived from it, the frontend-plugin reconciliation,
and the load/install/uninstall/enable/disable/settings actions, along
with the domain-only derivePluginBackendTypes and loadActivePluginThemes
helpers. selectedPluginId moves with the domain.

selectPlugin stays in the root store: it opens the plugin-detail tab via
the module-local createTab factory and moves with the panel/tab slice
when that domain is extracted (same caveat as openTunnelEditorTab).

Pure structural move — store shape, action signatures, and behavior are
unchanged; AppState now extends PluginsSlice. Full frontend suite, tsc,
eslint, and prettier all green.
